requirements.txt
Flask==1.0.2
Werkzeug==0.15.3
flask_cors==3.0.7 
gunicorn==19.9.0
validate_email==1.3
nose==1.3.7
requests==2.21.0
argparse==1.4.0
uuid==1.30
flask-swagger-ui==3.20.9
PyYaml==5.3.1